simulador-atuarial-individual/backend/pyproject.toml
[project]
name = "prevlab"
version = "0.1.0"
description = "PrevLab - Plataforma de Simulação Atuarial"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "numpy>=1.24.3",
    "pandas>=2.0.3",
    "scipy>=1.11.4",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "aiofiles>=23.2.1",
    "jinja2>=3.1.2",
    "sqlmodel>=0.0.14",
    "alembic>=1.13.0",
    "pymort>=0.2.0",
    "openpyxl>=3.1.0",
    "requests>=2.31.0",
    "reportlab>=4.4.4",
    "matplotlib>=3.10.6",
    "weasyprint>=66.0",
    "pillow>=11.3.0",
    "xlsxwriter>=3.2.9",
    "pytest>=8.4.1",
    "httpx>=0.28.1",
    "pytest-asyncio>=1.1.0",
    "websocket-client>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
select = ["E", "F", "I"]
ignore = []
line-length = 88

[tool.black]
line-length = 88
